Switch to gpgsm / PKCS#7 / CMS file format.
gpgsm is a newly added feature to duplicity which uses an alternate encryption
file format PKCS#7 / CMS. This is a standard encryption file format which is
also compatible with other tools (openssl smime).

This lets us enroll devices which cannot run duplicity into the Foulab
centralized backup system, such as the Wifi router (which runs FreshTomato with
a minimized root filesystem which has openssl but not much else), and manage
the encryption key using a single keyring.